Generate a data grid.

To visualise a model, it is very useful to be able to generate an evenly spaced grid of points from the data. data_grid helps you do this by wrapping around expand(). data_grid_ is the standard evaluation version for use in programming.

data_grid(data, ..., .model = NULL)

data_grid_(data, ..., .dots, .model = NULL)

Arguments

data

A data frame

...

Variables passed on to expand()

.model

A model. If supplied, any predictors needed for the model not present in ... will be filled in with "typical" values.

.dots

Used to work around non-standard evaluation.

See also

seq_range() for generating ranges from continuous variables.

Examples

require("modelr")
data_grid(mtcars, vs, am)

#> # A tibble: 4 × 2
#>      vs    am
#>   <dbl> <dbl>
#> 1     0     0
#> 2     0     1
#> 3     1     0
#> 4     1     1

# For continuous variables, seq_range is useful
data_grid(mtcars, mpg = seq_range(mpg, 10))

#> # A tibble: 10 × 1
#>         mpg
#>       <dbl>
#> 1  10.40000
#> 2  13.01111
#> 3  15.62222
#> 4  18.23333
#> 5  20.84444
#> 6  23.45556
#> 7  26.06667
#> 8  28.67778
#> 9  31.28889
#> 10 33.90000

# If you optionally supply a model, missing predictors will
# be filled in with typical values
mod <- lm(mpg ~ wt + cyl + vs, data = mtcars)
data_grid(mtcars, .model = mod)

#> # A tibble: 1 × 3
#>      wt   cyl    vs
#>   <dbl> <dbl> <dbl>
#> 1 3.325     6     0

data_grid(mtcars, cyl = seq_range(cyl, 9), .model = mod)

#> # A tibble: 9 × 3
#>     cyl    wt    vs
#>   <dbl> <dbl> <dbl>
#> 1   4.0 3.325     0
#> 2   4.5 3.325     0
#> 3   5.0 3.325     0
#> 4   5.5 3.325     0
#> 5   6.0 3.325     0
#> 6   6.5 3.325     0
#> 7   7.0 3.325     0
#> 8   7.5 3.325     0
#> 9   8.0 3.325     0