Filter rows with missing values

Remove rows with missing values in variables that are selected.

filter_na(.data, ..., .all = FALSE, .finite = FALSE)

filter_na_(.data, ..., .dots, .all = FALSE, .finite = FALSE)

Arguments

.data

A tbl

...

Comma separated list of unquoted expressions. Select variables which to test for missing values using the same semantics as select.

.all

If TRUE, then remove rows in which all columns are missing, else select rows in which any column has a missing value.

.finite

For numeric columns, filter using is.finite, which will drop rows with NaN, -Inf, and Inf values in addition to NA.

.dots

use filter_na_() to do standard evaluation. See vignette("nse") for details.

Value

An object of the same class as .data.